README.md
This Gallery app takes up image URLs from an ENDPOINT that gives us data in the form of JSON. It extracts URLs and other required information like caption and date, and display it into a Recycle-view. Further scrolling a added that is pictures can be viewed similar as that of any Gallery app just by scrolling.